diarization/requirements.txt
# Voice Diarization Pipeline — dependencies
# 
# PREFEROWANY INSTALL (uv):
#   uv sync
#
# Alternatywnie (pip):
#   pip install torch==2.4.1 torchaudio==2.4.1 --index-url https://download.pytorch.org/whl/cpu
#   pip install -r requirements.txt

# PyTorch CPU-only — PINOWANE wersje (pyannote.audio 3.3.x wymaga torchaudio <2.5)
# Jeśli instalujesz osobno:
#   pip install torch==2.4.1 torchaudio==2.4.1 --index-url https://download.pytorch.org/whl/cpu
torch==2.4.1
torchaudio==2.4.1

# Speaker diarization
pyannote.audio==3.3.2
huggingface_hub<0.25        # pyannote 3.3.x używa use_auth_token (usunięte w 0.25+)

# Voice activity detection
silero-vad

# Speaker embeddings (cross-recording matching)
speechbrain

# Audio manipulation
pydub

# Clustering & metrics
scikit-learn
numpy

# Data export
pandas

# Pretty terminal output
rich
